Time: 422 ms (17.49%), Space: 125 MB (50.41%) - LeetHub

+++ 0295-find-median-from-data-stream/0295-find-median-from-data-stream.java
@@ -16,7 +16,7 @@ class MedianFinder {
     }
 
     public double findMedian() {
-        if (even) return (double) (heapMin.peek() + heapMax.peek()) / 2.0;
+        if (even) return (heapMin.peek() + heapMax.peek()) / 2.0;
         return heapMax.peek();
     }
 }
